@@ -15,8 +15,10 @@
         }
     },
     "require": {
-        "php": "^5.6|^7.0",
-        "ext-iconv": "*"
+        "php": "^7.4|^8",
+        "ext-iconv": "*",
+        "ext-json": "*",
+        "symfony/polyfill-php80": "^1.26"
     },
     "require-dev": {
         "timber/timber": "^1.0",
@@ -24,6 +26,11 @@
         "roots/wordpress": "^5|^6",
         "twig/twig": "^2|^3",
         "cocur/slugify": "^3.1|^4.0"
+    },
+    "replace": {
+        "symfony/polyfill-php72": "*",
+        "symfony/polyfill-php73": "*",
+        "symfony/polyfill-php74": "*"
     },
     "suggest": {
         "cocur/slugify": "Provides ability to generate slugs if there is no transliteration is supported by iconv"
