@@ -7,31 +7,17 @@ namespace Flying\Wordpress;
  */
 class Autoloader
 {
-    /**
-     * @var string
-     */
-    private $root;
-    /**
-     * @var array
-     */
-    private $transformers;
+    private string $root;
+    private array $transformers;
 
-    /**
-     * @param string $root
-     * @param array $transformers
-     * @param bool $skipDefaultTransformers
-     * @throws \Exception
-     */
-    public function __construct($root, array $transformers = [], $skipDefaultTransformers = false)
+    public function __construct(string $root, array $transformers = [], bool $skipDefaultTransformers = false)
     {
         $root = rtrim(str_replace('\\', '/', $root), '/');
         if (!is_dir($root)) {
-            throw new \InvalidArgumentException('Given Wordpress root directory is not actually exists');
+            throw new \InvalidArgumentException('Given WordPress root directory does not actually exists');
         }
         $this->root = $root;
-        if (!empty(array_filter($transformers, function ($transformer) {
-            return !is_callable($transformer);
-        }))) {
+        if (!empty(array_filter($transformers, static fn($transformer) => !is_callable($transformer)))) {
             throw new \InvalidArgumentException('Class name transformers should be defined as array of callables');
         }
         if (!$skipDefaultTransformers) {
@@ -43,10 +29,8 @@ class Autoloader
 
     /**
      * Installs this class loader on the SPL autoload stack.
-     *
-     * @throws \Exception
      */
-    public function register()
+    public function register(): void
     {
         spl_autoload_register([$this, 'loadClass']);
     }
@@ -54,17 +38,15 @@ class Autoloader
     /**
      * Uninstalls this class loader from the SPL autoloader stack.
      */
-    public function unregister()
+    public function unregister(): void
     {
         spl_autoload_unregister([$this, 'loadClass']);
     }
 
     /**
      * Load given class
-     *
-     * @param string $class
      */
-    public function loadClass($class)
+    public function loadClass(string $class): void
     {
         $root = $this->getRoot();
         foreach ($this->transformers as $transformer) {
@@ -72,53 +54,37 @@ class Autoloader
             if (!is_string($path)) {
                 continue;
             }
-            if (file_exists($path)) {
-                /** @noinspection PhpIncludeInspection */
+            if (is_file($path)) {
                 include_once $path;
                 return;
             }
             $path = $root . '/' . $path;
-            if (file_exists($path)) {
-                /** @noinspection PhpIncludeInspection */
+            if (is_file($path)) {
                 include_once $path;
                 return;
             }
         }
     }
 
-    /**
-     * @return string
-     */
-    protected function getRoot()
+    protected function getRoot(): string
     {
         return $this->root;
     }
 
-    /**
-     * @return array
-     */
-    protected function getTransformers()
+    protected function getTransformers(): array
     {
         return $this->transformers;
     }
 
     /**
      * Get default class name transformation functions
-     *
-     * @return array
      */
-    protected function getDefaultTransformers()
+    protected function getDefaultTransformers(): array
     {
         return [
-            function ($class) {
-                return str_replace('_', '/', $class) . '.php';
-            },
-            function ($class) {
-                return strtolower(str_replace('_', '/', $class)) . '.php';
-            },
-            function ($class) {
-                return 'class-' . strtolower(str_replace('_', '-', $class)) . '.php';
-            },
+            fn(string $class): string => str_replace('_', '/', $class) . '.php',
+            fn(string $class): string => strtolower(str_replace('_', '/', $class)) . '.php',
+            fn(string $class): string => 'class-' . strtolower(str_replace('_', '-', $class)) . '.php',
         ];
     }
 }
