@@ -10,18 +10,18 @@ class CssUtils
      * @param mixed ...$args
      * @return string
      */
-    public static function classList(...$args)
+    public static function classList(...$args): string
     {
         $classes = [];
-        array_walk($args, function ($arg) use (&$classes) {
+        array_walk($args, static function ($arg) use (&$classes) {
             $parts = [];
             if (is_array($arg)) {
                 $parts = $arg;
             } elseif (is_string($arg)) {
                 $parts = explode(' ', $arg);
             }
-            array_walk_recursive($parts, function ($cl) use (&$classes) {
-                $cl = trim($cl);
+            array_walk_recursive($parts, static function ($cl) use (&$classes) {
+                $cl = trim((string)$cl);
                 if ($cl !== '' && !in_array($cl, $classes, true)) {
                     $classes[] = $cl;
                 }
@@ -36,7 +36,8 @@ class CssUtils
      * @param callable $modificator
      * @return string
      */
-    private static function updateClassList($classList, $class, $modificator) {
+    private static function updateClassList($classList, string $class, callable $modificator): string
+    {
         if (is_string($classList)) {
             $classList = explode(' ', $classList);
         }
@@ -53,9 +54,9 @@ class CssUtils
      * @param string $class
      * @return string
      */
-    public static function addClass($classList, $class)
+    public static function addClass($classList, string $class): string
     {
-        return self::updateClassList($classList, $class, function (array $classList, $class) {
+        return self::updateClassList($classList, $class, static function (array $classList, $class) {
             if (!in_array($class, $classList, true)) {
                 $classList[] = $class;
             }
@@ -70,12 +71,12 @@ class CssUtils
      * @param string $class
      * @return string
      */
-    public static function removeClass($classList, $class)
+    public static function removeClass($classList, string $class): string
     {
-        return self::updateClassList($classList, $class, function (array $classList, $class) {
-            return array_filter($classList, function ($v) use ($class) {
-                return $v !== $class;
-            });
-        });
+        return self::updateClassList(
+            $classList,
+            $class,
+            static fn(array $classList, $class) => array_filter($classList, static fn($v) => $v !== $class)
+        );
     }
 }
